corrige de bug mais pas fini

+++ src/main/java/mathis/simple_website_backend/repository/PeopleRepository.java
@@ -3,6 +3,6 @@ package mathis.simple_website_backend.repository;
 import mathis.simple_website_backend.models.People;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PeopleRepository extends JpaRepository<People, Integer> {
+public interface PeopleRepository extends JpaRepository<People, Long> {
     public People findPeopleByPrenom(String prenom);
 }
